+++ src/global.d.ts
@@ -3,6 +3,11 @@ export {};
 declare global {
   interface Window {
     OuraRingExtension?: unknown;
+    roamAlphaAPI?: {
+      constants?: {
+        corsAnywhereProxyUrl?: string;
+      };
+    };
   }
 }